今更だけどHomeControllerを3つに分けました 1.MypageController追加 2.BbsController追加 3.ManagementController追加 HomaController内のメソッドを上記３つのファイルへ転記しています。
