+++ app/views/bbs/bbs_cont2.scala.html
@@ -34,7 +34,7 @@
 			<center>
 				<h1>過去の事例</h1>
 				@for(date <- dateList) {
-					<a class="popup-iframe" href="@routes.ModelController.bbs_cont4(date)">● @date TOP10</a></br>
+					<a class="popup-iframe" href="@routes.BbsController.bbs_cont4(date)">● @date TOP10</a></br>
 				}
 			</center>
 		</div>
