+++ app/views/card.scala.html
@@ -1,144 +1,116 @@
 <!DOCTYPE html>
+
 <html lang="ja">
-<head>
-    <meta charset="utf-8">
-           <!-- jQuery -->
-<style>
-     .myButtonI {
-	-moz-box-shadow:inset 0px 1px 0px 0px #97c4fe;
-	-webkit-box-shadow:inset 0px 1px 0px 0px #97c4fe;
-	box-shadow:inset 0px 1px 0px 0px #97c4fe;
-	background:-webkit-gradient(linear, left top, left bottom, color-stop(0.05, #3d94f6), color-stop(1, #1e62d0));
-	background:-moz-linear-gradient(top, #3d94f6 5%, #1e62d0 100%);
-	background:-webkit-linear-gradient(top, #3d94f6 5%, #1e62d0 100%);
-	background:-o-linear-gradient(top, #3d94f6 5%, #1e62d0 100%);
-	background:-ms-linear-gradient(top, #3d94f6 5%, #1e62d0 100%);
-	background:linear-gradient(to bottom, #3d94f6 5%, #1e62d0 100%);
-	filter:progid:DXImageTransform.Microsoft.gradient(startColorstr='#3d94f6', endColorstr='#1e62d0',GradientType=0);
-	background-color:#3d94f6;
-	-moz-border-radius:6px;
-	-webkit-border-radius:6px;
-	border-radius:6px;
-	border:1px solid #337fed;
-	display:inline-block;
-	cursor:pointer;
-	color:#ffffff;
-	font-family:Arial;
-	font-size:15px;
-	font-weight:bold;
-	padding:6px 24px;
-	text-decoration:none;
-	text-shadow:0px 1px 0px #1570cd;
-}
-.myButtonI:hover {
-	background:-webkit-gradient(linear, left top, left bottom, color-stop(0.05, #1e62d0), color-stop(1, #3d94f6));
-	background:-moz-linear-gradient(top, #1e62d0 5%, #3d94f6 100%);
-	background:-webkit-linear-gradient(top, #1e62d0 5%, #3d94f6 100%);
-	background:-o-linear-gradient(top, #1e62d0 5%, #3d94f6 100%);
-	background:-ms-linear-gradient(top, #1e62d0 5%, #3d94f6 100%);
-	background:linear-gradient(to bottom, #1e62d0 5%, #3d94f6 100%);
-	filter:progid:DXImageTransform.Microsoft.gradient(startColorstr='#1e62d0', endColorstr='#3d94f6',GradientType=0);
-	background-color:#1e62d0;
-}
-.myButtonI:active {
-	position:relative;
-	top:1px;
-}
-    div {
-        font-family: cursive;
-
-        margin: 50px;
-
-    }
-
-    sample2Wrap{
-
-}
-.sample2{
-	width: 100%;
-	height: 100%;
-	background: #c1f5bc;
-	margin: 0 auto;
-    margin-top: 35px;
-}
-    .sample3{
-
-    font-size: 20px;
-	padding: 50px;
-}
-
- .dl2 dl{
-   width : 700px;
-}
-
-.dl2 dt{
-  font-size : 15px;
-  font-weight : bold;
-
-  margin-bottom : 10px;
-
-  border-bottom-width : 1px;
-  border-bottom-style : solid;
-  border-bottom-color : #666666;
-
-  border-left-width : 7px;
-  border-left-style : solid;
-  border-left-color : #666666;
-
-  padding-top : 2px;
-  padding-left : 8px;
-  padding-right : 8px;
-  padding-bottom : 2px;
-}
-
-.dl2 dd{
-  margin-right: 10px;
-  padding: 10px;
-  font-size : 80%;
-  line-height : 1.8;
-  margin-bottom : 45px;
-
-
-  border-bottom-width : 1px;
-  border-bottom-style : solid;
-  border-bottom-color : #999999;
-
-  padding-left : 30px;
-  padding-right : 15px;
-
-}
-</style>
-
-    </head>
-
-<body>
-  <div class="sample2Wrap">
-	<div class="sample2">
-        <div>
-            <br>
-          <center>  <h1>～ THANK YOU ～</h1></center>
-
-    <p>●TO　: 金城</p>
-
-
-
-   <center> <p>●TITLE　:  説明書き説明書き説明書き説明書き</p></center>
-
-
-
-        <dl class="dl2">
-            <dt><a>●HELP   &nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp &nbsp&nbsp&nbsp&nbsp●CATEGORY:</a> </dt>
-        <dd>サンプルサンプルサンプルサンプルサンプル・・・・・・</dd>
-            <dt>●MESSAGE</dt>
-            <dd>サンプルサンプルサンプルサンプルサンプル・・・・・・</dd>
-        </dl>
-
-
-        <p Align="right">●FROM : 金村 <br> <br><input class="myButtonI" type="submit" style="width: 100px; height: 30px;" value="いいね!!" /></p>
-
-                </div>
-      </div>
-      </div>
-
-</body>
+	<head>
+		<meta charset="utf-8">
+		<!-- jQuery -->
+		<style>
+			.myButtonI {
+				-moz-box-shadow:inset 0px 1px 0px 0px #97c4fe;
+				-webkit-box-shadow:inset 0px 1px 0px 0px #97c4fe;
+				box-shadow:inset 0px 1px 0px 0px #97c4fe;
+				background:-webkit-gradient(linear, left top, left bottom, color-stop(0.05, #3d94f6), color-stop(1, #1e62d0));
+				background:-moz-linear-gradient(top, #3d94f6 5%, #1e62d0 100%);
+				background:-webkit-linear-gradient(top, #3d94f6 5%, #1e62d0 100%);
+				background:-o-linear-gradient(top, #3d94f6 5%, #1e62d0 100%);
+				background:-ms-linear-gradient(top, #3d94f6 5%, #1e62d0 100%);
+				background:linear-gradient(to bottom, #3d94f6 5%, #1e62d0 100%);
+				filter:progid:DXImageTransform.Microsoft.gradient(startColorstr='#3d94f6', endColorstr='#1e62d0',GradientType=0);
+				background-color:#3d94f6;
+				-moz-border-radius:6px;
+				-webkit-border-radius:6px;
+				border-radius:6px;
+				border:1px solid #337fed;
+				display:inline-block;
+				cursor:pointer;
+				color:#ffffff;
+				font-family:Arial;
+				font-size:15px;
+				font-weight:bold;
+				padding:6px 24px;
+				text-decoration:none;
+				text-shadow:0px 1px 0px #1570cd;
+			}
+			.myButtonI:hover {
+				background:-webkit-gradient(linear, left top, left bottom, color-stop(0.05, #1e62d0), color-stop(1, #3d94f6));
+				background:-moz-linear-gradient(top, #1e62d0 5%, #3d94f6 100%);
+				background:-webkit-linear-gradient(top, #1e62d0 5%, #3d94f6 100%);
+				background:-o-linear-gradient(top, #1e62d0 5%, #3d94f6 100%);
+				background:-ms-linear-gradient(top, #1e62d0 5%, #3d94f6 100%);
+				background:linear-gradient(to bottom, #1e62d0 5%, #3d94f6 100%);
+				filter:progid:DXImageTransform.Microsoft.gradient(startColorstr='#1e62d0', endColorstr='#3d94f6',GradientType=0);
+				background-color:#1e62d0;
+			}
+			.myButtonI:active {
+				position:relative;
+				top:1px;
+			}
+			div {
+				font-family: cursive;
+				margin: 50px;
+			}
+			sample2Wrap {
+			}
+			.sample2 {
+				width: 100%;
+				height: 100%;
+				background: #c1f5bc;
+				margin: 0 auto;
+				margin-top: 35px;
+			}
+			.sample3 {
+				font-size: 20px;
+				padding: 50px;
+			}
+			.dl2 dl {
+				width : 700px;
+			}
+			.dl2 dt {
+				font-size : 15px;
+				font-weight : bold;
+				margin-bottom : 10px;
+				border-bottom-width : 1px;
+				border-bottom-style : solid;
+				border-bottom-color : #666666;
+				border-left-width : 7px;
+				border-left-style : solid;
+				border-left-color : #666666;
+				padding-top : 2px;
+				padding-left : 8px;
+				padding-right : 8px;
+				padding-bottom : 2px;
+			}
+			.dl2 dd {
+				margin-right: 10px;
+				padding: 10px;
+				font-size : 80%;
+				line-height : 1.8;
+				margin-bottom : 45px;
+				border-bottom-width : 1px;
+				border-bottom-style : solid;
+				border-bottom-color : #999999;
+				padding-left : 30px;
+				padding-right : 15px;
+			}
+		</style>
+	</head>
+	<body>
+		<div class="sample2Wrap">
+			<div class="sample2">
+				<div>
+					<br>
+					<center>  <h1>～ THANK YOU ～</h1></center>
+					<p>●TO　: 金城</p>
+					<center><p>●TITLE　:  説明書き説明書き説明書き説明書き</p></center>
+					<dl class="dl2">
+						<dt><a>●HELP   &nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp &nbsp&nbsp&nbsp&nbsp●CATEGORY:</a> </dt>
+						<dd>サンプルサンプルサンプルサンプルサンプル・・・・・・</dd>
+						<dt>●MESSAGE</dt>
+						<dd>サンプルサンプルサンプルサンプルサンプル・・・・・・</dd>
+					</dl>
+					<p Align="right">●FROM : 金村 <br><br><input class="myButtonI" type="submit" style="width: 100px; height: 30px;" value="いいね!!" /></p>
+				</div>
+			</div>
+		</div>
+	</body>
 </html>
