+++ app/views/main_frame.scala.html
@@ -49,9 +49,9 @@
 		<!-- グローバルナビ -->
 		<nav id="global_navi">
 			<ul>
-				<li><a href="@routes.HomeController.mypage_main()">マイページ</a></li>
-				<li><a href="@routes.ModelController.bbs_main()">掲示板</a></li>
-				<li><a href="@routes.HomeController.management_main()">管理者設定</a></li>
+				<li><a href="@routes.MypageController.mypage_main()">マイページ</a></li>
+				<li><a href="@routes.BbsController.bbs_main()">掲示板</a></li>
+				<li><a href="@routes.ManagementController.management_main()">管理者設定</a></li>
 				<li><a href="@routes.AuthController.logout()">ログアウト</a></li>
 			</ul>
 		</nav>
@@ -66,9 +66,9 @@
 		<footer>
 			<div id="footer_nav">
 				<ul>
-					<li><a href="@routes.HomeController.mypage_main()">マイページ</a></li>
-					<li><a href="@routes.ModelController.bbs_main()">掲示板</a></li>
-					<li><a href="@routes.HomeController.management_main()">管理者画面</a></li>
+					<li><a href="@routes.MypageController.mypage_main()">マイページ</a></li>
+					<li><a href="@routes.BbsController.bbs_main()">掲示板</a></li>
+					<li><a href="@routes.ManagementController.management_main()">管理者画面</a></li>
 					<li><a href="@routes.AuthController.logout()">ログアウト</a>
 				</ul>
 			</div>
