+++ app/views/mypage/mypage_main.scala.html
@@ -29,13 +29,13 @@
 		<div id="main">
 			<!--表などを表示するコンテンツ部分-->
 			<div id="cont1">
-				<iframe src="@routes.HomeController.mypage_cont1()"width="100%" height="353px"frameborder="0"></iframe>
+				<iframe src="@routes.MypageController.mypage_cont1()"width="100%" height="353px"frameborder="0"></iframe>
 			</div>
 			<div id="cont2">
-				<iframe src="@routes.HomeController.mypage_cont2()"width="100%" height="353px"frameborder="0"></iframe>
+				<iframe src="@routes.MypageController.mypage_cont2()"width="100%" height="353px"frameborder="0"></iframe>
 			</div>
 			<div id="cont3">
-				<iframe src="@routes.HomeController.mypage_cont3()"width="100%" height="353px"frameborder="0"></iframe>
+				<iframe src="@routes.MypageController.mypage_cont3()"width="100%" height="353px"frameborder="0"></iframe>
 			</div>
 			<!-- jQueryを使ったタブのスクリプト処理 -->
 			<script>
