+++ app/views/system_main.scala.html
@@ -63,9 +63,9 @@
 		<footer>
 			<div id="footer_nav">
 				<ul>
-					<li><a href="@routes.HomeController.mypage_main()">マイページ</a></li>
-					<li><a href="@routes.ModelController.bbs_main()">掲示板</a></li>
-					<li><a href="@routes.HomeController.management_main()">管理者画面</a></li>
+					<li><a href="@routes.MypageController.mypage_main()">マイページ</a></li>
+					<li><a href="@routes.BbsController.bbs_main()">掲示板</a></li>
+					<li><a href="@routes.ManagementController.management_main()">管理者画面</a></li>
 					<li><a href="@routes.AuthController.logout()">ログアウト</a>
 				</ul>
 			</div>
